Move config and templates to scripts folder.
