+++ bin/styleguidist
@@ -7,7 +7,7 @@ const chalk = require('chalk');
 
 const argv = minimist(process.argv.slice(2));
 
-const config = require('./../src/utils/config');
+const config = require('./../scripts/config');
 config.initialize(argv);
 
 switch (argv._[0]) {
