+++ scripts/templates/DefaultExample.md
